Module sw8s_rust_lib::missions::action

Structs

ActionChain

ActionConcurrent

ActionConcurrentSplit

ActionConditional

An action that runs one of two actions depending on if its conditional reference is true or false.

ActionDataConditional

An action that runs one of two actions depending on if its conditional reference is valid or not, passing processed output to the first child and unprocessed output to the second child.

ActionParallel

ActionSelect

ActionSequence

ActionUntil

An action that tries count times for a success

ActionWhile

An action that runs while true

DualAction

Run two actions at once, and only exit when all actions have exited.

FirstValid

Return first valid response from block of actions.

RaceAction

Action that runs two actions at the same time and exits both when one exits

TupleSecond

Get second arg in action output

Traits

Action

A trait for an action that can be executed.

ActionExec

A trait for an action that can be executed.

ActionIgnoredGeneric

ActionMod

An action that can be modified at runtime.

Functions

wrap_action

Produces a new function that puts wrapper around wrapee.