sw8s_rust_lib
Module vision

Modules

buoy

buoy_model

gate

gate_poles

image_prep

nn_cv2

octagon

path

pca

yolo_model

Structs

Angle2D

Holds x, y, and angle offset of object in frame

DrawRect2d

MatWrapper

Allows Mat to be shared across threads for async. The C pointer is perfectly safe to share between threads, Rust just defaults to not giving any pointer Send/Sync so we have to use this wrapper pattern.

Offset2D

Holds x and y offset of object in frame

VecMatWrapper

Allows Vector<Mat> to be shared across threads for async. The C pointer is perfectly safe to share between threads, Rust just defaults to not giving any pointer Send/Sync so we have to use this wrapper pattern.

VisualDetection

Traits

Draw

RelPos

RelPosAngle

VisualDetector

Functions

mean