Struct sw8s_rust_lib::vision::VecMatWrapper

pub struct VecMatWrapper(pub Vector<Mat>);

Allows Vector<Mat> to be shared across threads for async. The C pointer is perfectly safe to share between threads, Rust just defaults to not giving any pointer Send/Sync so we have to use this wrapper pattern.

Tuple Fields
0: Vector<Mat>