Trait sw8s_rust_lib::missions::action::Action

pub trait Action {
    // Provided method
    fn dot_string(&self, _parent: &str) -> DotString { ... }
}

A trait for an action that can be executed.

Functions returning Actions/ActionExec should be written with -> impl ActionExec + '_ and using Con for the context generic. This allows the build script to strip out the context and create a mirrored version returning -> impl Action +'_ for graphing.

Provided Methods

fn dot_string(&self, _parent: &str) -> DotString

Represent this node in dot (graphviz) notation