Module sw8s_rust_lib::vision

Modules

buoy

gate

gate_poles

image_prep

nn_cv2

path

pca

yolo_model

Structs

Angle2D

Holds x, y, and angle offset of object in frame

DrawRect2d

Offset2D

Holds x and y offset of object in frame

VisualDetection

Traits

Draw

RelPos

RelPosAngle

VisualDetector

Functions

mean